+++ src/components.d.ts
@@ -9,6 +9,7 @@ import { TecnologiaTheme } from "./components/interfaces";
 import { ButtonPosition } from "./components/bottom-bar/bottom-bar.model";
 import { TecStatus } from "./models/status.model";
 import { TecButtonColor, TecButtonIconMode, TecButtonMode, TecButtonSize } from "./components/tec-button/tec-button.model";
+import { TecSize } from "./models/size.model";
 export namespace Components {
     interface TecBottomBar {
         "buttonPosition": ButtonPosition;
@@ -64,6 +65,18 @@ export namespace Components {
         "status": TecStatus;
         "theme": TecnologiaTheme;
     }
+    interface TecModal {
+        "blockBodyScroll": boolean;
+        "closeOnEscape": boolean;
+        "dismissOnBackdrop": boolean;
+        "fullWidth": boolean;
+        "modalTitle": string;
+        "opened": boolean;
+        "responsive": boolean;
+        "showCloseIcon": boolean;
+        "size": TecSize;
+        "theme": TecnologiaTheme;
+    }
     interface TecProductHeader {
         /**
           * Use to make a bar fixed on top
@@ -114,6 +127,12 @@ declare global {
         prototype: HTMLTecButtonElement;
         new (): HTMLTecButtonElement;
     };
+    interface HTMLTecModalElement extends Components.TecModal, HTMLStencilElement {
+    }
+    var HTMLTecModalElement: {
+        prototype: HTMLTecModalElement;
+        new (): HTMLTecModalElement;
+    };
     interface HTMLTecProductHeaderElement extends Components.TecProductHeader, HTMLStencilElement {
     }
     var HTMLTecProductHeaderElement: {
@@ -129,6 +148,7 @@ declare global {
     interface HTMLElementTagNameMap {
         "tec-bottom-bar": HTMLTecBottomBarElement;
         "tec-button": HTMLTecButtonElement;
+        "tec-modal": HTMLTecModalElement;
         "tec-product-header": HTMLTecProductHeaderElement;
     }
 }
@@ -191,6 +211,19 @@ declare namespace LocalJSX {
         "status"?: TecStatus;
         "theme"?: TecnologiaTheme;
     }
+    interface TecModal {
+        "blockBodyScroll"?: boolean;
+        "closeOnEscape"?: boolean;
+        "dismissOnBackdrop"?: boolean;
+        "fullWidth"?: boolean;
+        "modalTitle"?: string;
+        "onHidden"?: (event: CustomEvent<UIEvent>) => void;
+        "opened"?: boolean;
+        "responsive"?: boolean;
+        "showCloseIcon"?: boolean;
+        "size"?: TecSize;
+        "theme"?: TecnologiaTheme;
+    }
     interface TecProductHeader {
         /**
           * Emitted when the title was clicked
@@ -240,6 +273,7 @@ declare namespace LocalJSX {
     interface IntrinsicElements {
         "tec-bottom-bar": TecBottomBar;
         "tec-button": TecButton;
+        "tec-modal": TecModal;
         "tec-product-header": TecProductHeader;
     }
 }
@@ -249,6 +283,7 @@ declare module "@stencil/core" {
         interface IntrinsicElements {
             "tec-bottom-bar": LocalJSX.TecBottomBar & JSXBase.HTMLAttributes<HTMLTecBottomBarElement>;
             "tec-button": LocalJSX.TecButton & JSXBase.HTMLAttributes<HTMLTecButtonElement>;
+            "tec-modal": LocalJSX.TecModal & JSXBase.HTMLAttributes<HTMLTecModalElement>;
             "tec-product-header": LocalJSX.TecProductHeader & JSXBase.HTMLAttributes<HTMLTecProductHeaderElement>;
         }
     }