+++ src/components/tec-modal/test/tec-modal.e2e.ts
@@ -1,0 +1,11 @@
+import { newE2EPage } from '@stencil/core/testing';
+
+describe('tec-modal', () => {
+  it('renders', async () => {
+    const page = await newE2EPage();
+    await page.setContent('<tec-modal></tec-modal>');
+
+    const element = await page.find('tec-modal');
+    expect(element).toHaveClass('hydrated');
+  });
+});