+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release notes
 
+## 0.5.0
+
+### Features
+
+- Created `modal` component.
+
+### Fixes
+
+- Set `reflect: true` on button component.
+
 ## 0.4.1
 
 ### Fixes